feat(foundation): add progress bar component
